refactor: dev pull and refactor
