--- src/main/java/com/planme/main/SecurityConfig.java
@@ -1,22 +1,0 @@
-package com.planme.main;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-//401 방지로 잠시 사용하는 클래스입니다
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .authorizeRequests(authorize -> authorize
-                        .anyRequest().permitAll())// 모든 요청에 대해 인증 없이 접근을 허용
-                .csrf(csrf -> csrf.disable());
-        return http.build();
-    }
-}
