+++ src/main/java/com/planme/main/service/focusService/FocusQueryServiceImpl.java
@@ -1,4 +1,4 @@
-package com.planme.main.service;
+package com.planme.main.service.focusService;
 
 import com.planme.main.apiPayload.code.status.ErrorStatus;
 import com.planme.main.apiPayload.exception.handler.FocusHandler;
@@ -7,7 +7,6 @@ import com.planme.main.repository.FocusRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
