+++ src/main/java/com/planme/main/web/dto/FocusDTO/FocusRequestDTO.java
@@ -1,4 +1,4 @@
-package com.planme.main.web.dto;
+package com.planme.main.web.dto.FocusDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
